Add spacing between column name and type [ME-8]
Also, italicize the type and remove the wrapping parentheses. The
lighter gray color and italics are enough to set it apart visually; the
parens are just clutter.

Also, render model name with bold font, rather than italics.

Also, remove "Associations" heading. I think it's relatively obvious
what is in that section, and I think it looks a bit cleaner without the
label. Also, this is consistent with our not having a "Columns" label
for that section.