Add spacing between column name and type [ME-8] (#30)
Also, italicize the type and remove the wrapping parentheses. The
lighter gray color and italics are enough to set it apart visually; the
parens are just clutter.

Also, render model name with bold font, rather than italics.

Also, remove "Associations" heading. I think it's relatively obvious
what is in that section, and I think it looks a bit cleaner without the
label. Also, this is consistent with our not having a "Columns" label
for that section.

| Before | After |
| - | - |
|
![image](https://github.com/user-attachments/assets/f0907114-d858-4bd2-869f-37e775c7e4dc)
|
![image](https://github.com/user-attachments/assets/65bd3fc8-3f86-4f73-b5a2-611fe5837fbf)
|